requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
pydantic==2.10.6
python-dotenv==1.0.0
google-genai==1.7.0
gunicorn==23.0.0
python-multipart==0.0.20
